+++ packages/s2-react/playground/config.ts
@@ -77,12 +77,15 @@ export const pivotSheetDataCfgForCompactMode = customMerge(pivotSheetDataCfg, {
 });
 
 export const s2Options: SheetComponentOptions = {
-  debug: true,
+  debug: false,
   width: 600,
   height: 400,
-  showSeriesNumber: false,
+  frozenFirstRow: false,
+  showSeriesNumber: true,
   interaction: {
     enableCopy: true,
+    copyWithHeader: true,
+    copyWithFormat: true,
     // 防止 mac 触摸板横向滚动触发浏览器返回, 和移动端下拉刷新
     overscrollBehavior: 'none',
     brushSelection: {
@@ -140,7 +143,7 @@ export const s2Options: SheetComponentOptions = {
     },
     cellCfg: {
       height: 50,
-      width: 200,
+      width: 100,
     },
   },
 };